perf(autopilot): skip duplicate note candidates within one run
run_notes_loop keyed only against prior runs, so the same NEW job surfaced
under two URL variants in one run bought an AI note for each; the later
merge_found_jobs collapsed them and discarded one - waste against the
ASSISTANT_NOTES_MAX ceiling. Track canonical keys seen this run and skip the
second variant (before charge_daily, so it also spares the shared ceiling).

The oversized autopilot_helpers inline test module is moved to a sibling
tests.rs so mod.rs stays under the R8 LOC cap; no test behavior changes.

Source: #813 follow-up.
